[Refactor] 일기 스케줄러 구조 수정
